Migrated to EGLFS, Qt5.4 and QOpenglWidget

+++ openpandora/pandora_develop_environment/pandora_develop_environment.sh
@@ -1,11 +1,25 @@
 #必要ツール
 sudo dpkg --add-architecture i386
-sudo apt-get install ia32-libs libssl-dev libcurl4-openssl-dev libgpgme11-dev libtool build-essential libglib2.0-dev xsltproc squashfs-tools libncurses5:i386
+sudo apt-get install libssl-dev libcurl4-openssl-dev libgpgme11-dev libtool build-essential libglib2.0-dev xsltproc squashfs-tools libncurses5:i386
 sudo apt-get build-dep gdb
 
 #カレントディレクトリ
 INSTALLER_DIR=$PWD                                                  
 
+#SDKディレクトリ
+export PNDSDK=$HOME/pandora-dev/arm-2010q1
+
+#SDKフォルダ内のホームフォルダ
+export SDKHOME=$PNDSDK/home/$USER
+
+
+#初回起動時に環境全体を構築する
+if [ ! -d $PNDSDK ]; then
+FIRSTRUN=1
+else
+FIRSTRUN=0
+fi
+	
 #sdk_installer_openpandora_toolchain.sh中の
 #pandora_configure.sh
 #CPPFLAGSとCXXFLAGSを交換
@@ -15,22 +29,20 @@ INSTALLER_DIR=$PWD
 # s#pluginsdir=\/.*#pluginsdir='$PNDSDK/usr/libgstreamer-0.10'#;
 #を追加しておく
 #(予め編集しておく)
+if [ $FIRSTRUN -eq 1 ]; then 
 ./sdk_installer_openpandora_toolchain.sh
-export PNDSDK=$HOME/pandora-dev/arm-2010q1
 ln -snf $PNDSDK $HOME/pandora-dev/arm
-
-#cmathの不具合対策
-find $PNDSDK/arm-none-linux-gnueabi/include/c++ -name cmath | xargs sed -i -e "s|  using ::\([a-z0-9]\+\)l;|  //using ::\1l;|"
-find $PNDSDK/arm-none-linux-gnueabi/include/c++ -name cmath | xargs sed -i -e "s|  //using ::ceil;|  using ::ceil;|"
-
-
-#SDKフォルダ内のホームフォルダ
-export SDKHOME=$PNDSDK/home/$USER
+fi
 
 #pkgconfigを設定（プロジェクト設定にも追加？）
 export PKG_CONFIG_LIBDIR=$PNDSDK/usr/lib/pkgconfig:$PNDSDK/usr/share/pkgconfig
 export PKG_CONFIG_PATH=$PKG_CONFIG_LIBDIR
 export PKG_CONFIG_SYSROOT_DIR=/
+
+if [ $FIRSTRUN -eq 1 ]; then 
+#cmathの不具合対策
+find $PNDSDK/arm-none-linux-gnueabi/include/c++ -name cmath | xargs sed -i -e "s|  using ::\([a-z0-9]\+\)l;|  //using ::\1l;|"
+find $PNDSDK/arm-none-linux-gnueabi/include/c++ -name cmath | xargs sed -i -e "s|  //using ::ceil;|  using ::ceil;|"
 
 #libtoolはクロスコンパイルで問題が出るので関連ファイルを消しておく
 rm $PNDSDK/usr/lib/*.la
@@ -75,27 +87,49 @@ cd libX11-1.3.6
 ../../../../sdk_utils/pandora_configure.sh --prefix=$PNDSDK/usr --disable-malloc0returnsnull
 make -j3 && make install
 
-#Qt
-cd $SDKHOME
-QT_MAJOR_VER=5.3
-QT_VER=$QT_MAJOR_VER.2
-wget -c http://download.qt-project.org/official_releases/qt/$QT_MAJOR_VER/$QT_VER/single/qt-everywhere-opensource-src-$QT_VER.tar.gz -P $INSTALLER_DIR/tmp
-tar xf $INSTALLER_DIR/tmp/qt-everywhere-opensource-src-$QT_VER.tar.gz
-cd qt-everywhere-opensource-src-$QT_VER
+fi #if [ $FIRSTRUN -eq 1 ] 
 
-#mkspecをコピー
-cp -rf $INSTALLER_DIR/linux-pandora-g++ qtbase/mkspecs
-sed -i -e "s|\$\$PNDSDK|$PNDSDK|" qtbase/mkspecs/linux-pandora-g++/qmake.conf
+
+#Qt
+#インストールに使用するフォルダの名前。「qt5」という名前にしてはならない。
+QT_INSTALLNAME=qt540-beta-release
+
+QT_MAJOR_VER=5.4
+QT_VER=$QT_MAJOR_VER.0
+QT_FULL_VER=$QT_VER-beta
+QT_RELEASE=development_releases
+#QT_RELEASE=official_releases
+QT_SOURCE_NAME=qt-everywhere-opensource-src-$QT_FULL_VER
+
+cd $SDKHOME
+rm -rf $QT_INSTALLNAME
+rm -rf $PNDSDK/usr/local/$QT_INSTALLNAME
+mkdir $QT_INSTALLNAME
+
+wget -c http://download.qt-project.org/$QT_RELEASE/qt/$QT_MAJOR_VER/$QT_FULL_VER/single/$QT_SOURCE_NAME.tar.xz -P $INSTALLER_DIR/tmp
+
+if [ ! -d $QT_SOURCE_NAME ]; then
+tar Jxf $INSTALLER_DIR/tmp/$QT_SOURCE_NAME.tar.xz
+pushd $QT_SOURCE_NAME
 
 #ALSAで音が出ない問題に対処するパッチを当てる
 patch -p1  < $INSTALLER_DIR/qtmultimedia53.patch
 #カメラ関係でビルドが通らないのを回避
 rm -rf qtmultimedia/config.tests/gstreamer_encodingprofiles
 
-cd qtbase
+popd
+fi #if [ ! -d $QT_SOURCE_NAME ]; then
+
+#mkspecをコピー
+cd $QT_SOURCE_NAME
+cp -rf $INSTALLER_DIR/linux-pandora-g++ qtbase/mkspecs
+sed -i -e "s|\$\$PNDSDK|$PNDSDK|" qtbase/mkspecs/linux-pandora-g++/qmake.conf
+
+cd $SDKHOME/$QT_INSTALLNAME
 #make confclean -j3
-cd ..
-./configure -opensource -confirm-license -prefix $PNDSDK/usr -headerdir $PNDSDK/usr/include/qt5 -xplatform linux-pandora-g++ -static -opengl desktop -c++11 -qt-xcb -no-icu -no-pulseaudio -no-sql-sqlite -nomake examples -skip qtwebkit-examples -skip qtlocation -silent
+../$QT_SOURCE_NAME/configure -opensource -confirm-license -prefix $PNDSDK/usr/local/$QT_INSTALLNAME -xplatform linux-pandora-g++ -static -qreal float -opengl es2 -c++11 -qpa eglfs -no-xcb -no-icu -no-pulseaudio -no-sql-sqlite -nomake examples -skip qtwebkit-examples -skip qtlocation -continue -silent
+
 #echo "Hit Enter.";read Wait
 make -j3 && make install
+ln -snf $PNDSDK/usr/local/$QT_INSTALLNAME $PNDSDK/usr/local/qt5
